Renamed the files
1. Renamed the files to shorten the length
2. Changed the make files to accommodate the name change
3. Changed the include file names in driver/application
