+++ kmod/Makefile
@@ -79,7 +79,7 @@ DWC_ETH_QOS-y += DWC_ETH_QOS_dev.o \
 DWC_ETH_QOS-$(CONFIG_DEBUGFS_OBJ) += DWC_ETH_QOS_debug_operation.o
 DWC_ETH_QOS-$(CONFIG_PGTEST_OBJ) += DWC_ETH_QOS_pktgen.o
 DWC_ETH_QOS-$(CONFIG_PTPSUPPORT_OBJ) += DWC_ETH_QOS_ptp.o
-DWC_ETH_QOS-$(CONFIG_UMODEDRV_OBJ) += DWC_ETH_QOS_avb.o
+DWC_ETH_QOS-$(CONFIG_UMODEDRV_OBJ) += dwcqos_avb.o
 
 all:
 	make -C $(KERNELDIR) M=$(PWD) modules
