+++ umod/Makefile
@@ -1,7 +1,7 @@
-OBJS=DWC_ETH_QOS_drv.o
-OBJS1=DWC_ETH_QOS_dev.o
-INCL=DWC_ETH_QOS_yheader.h DWC_ETH_QOS_yregacc.h DWC_ETH_QOS_ycommon.h DWC_ETH_QOS_dev.h
-AVBLIB=libDWC_ETH_QOS_drv.a
+OBJS=dwcqos_drv.o
+OBJS1=dwcqos_dev.o
+INCL=dwcqos.h dwcqos_reg.h dwcqos_common.h dwcqos_dev.h
+AVBLIB=libdwcqos_drv.a
 
 all: $(AVBLIB)
 
@@ -9,11 +9,11 @@ $(AVBLIB):$(OBJS) $(OBJS1)
 	rm -f $@
 	ar -cvq $@ $(OBJS) $(OBJS1)
 
-$(OBJS): DWC_ETH_QOS_drv.c $(INCL)
-	gcc -c DWC_ETH_QOS_drv.c
+$(OBJS): dwcqos_drv.c $(INCL)
+	gcc -c dwcqos_drv.c
 
-$(OBJS1): DWC_ETH_QOS_dev.c $(INCL)
-	gcc -c DWC_ETH_QOS_dev.c
+$(OBJS1): dwcqos_dev.c $(INCL)
+	gcc -c dwcqos_dev.c
 
 clean:
 	rm -f `find . -name "*~" -o -name "*.[oa]" -o -name "\#*\#" -o -name TAGS -o -name core -o -name "*.orig"`
