+++ umod/dwcqos_common.h
